# docs/mess_openapi.yaml
...
openapi: "3.1.1"
info:
  title: Mess Management System API
  version: "2"
  summary: API Documentation for the Mess Management System
  description: |
    Please see the [tutorial on using MMS APIs](https://gitlab.iiit.ac.in/messcom/backend/-/blob/main/docs/tutorial.md).

    ## Rate Limits

    There are rudimentary rate limits put in place for every requests of **420 times a second**. There are unfortunately
    no `X-RateLimit-*` headers sent with the request as of the time of writing, but if a request is rate limited, a
    `429 Too Many Requests` is sent, with the JSON body being the typical error response with the code "rate-limited".

    ## Changes in v2
     
    - `/api/mess/minimal` is removed and replaced by `/api/mess/info` and no longer returns rates
    - `/api/mess/menus` endpoint no longer requires a mess ID, it returns all the menus
  contact:
    name: Mess Portal Admins
    email: messportal.admins@students.iiit.ac.in
  license:
    name: GNU Affero General Public License 3.0
    identifier: AGPL-3.0
servers:
  - url: /api
    description: This server
  - url: https://mess.iiit.ac.in/api
    description: Production server
...
paths:
...
  /mess/info:
    get:
      summary: Get information about all messes
      description: |
        **New in v2:** Does not return rates anymore
      tags:
        - Mess
      responses:
        '200':
          description: The information
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MessInfo"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError" 
...
  /mess/rates:
    get:
      summary: Get mess rates for a meal
      tags:
        - Mess
      parameters:
        - name: meal
          in: query
          description: The name of the meal.
          required: true
          schema:
            $ref: "#/components/schemas/Meal"
        - name: on
          in: query
          description: Meal date for getting the rate. Defaults to today's date.
          required: false
          schema:
            type: string
            format: date
            example: "1998-09-02"
      responses:
        '200':
          description: |
            The rates of each mess for the given meal on the given date. Note that
            each property will be a category, and the array of items will contain
            the rate and the day if a specific day's rate is applicable.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    propertyNames:
                      $ref: "#/components/schemas/Category"
                    additionalProperties:
                      type: array
                      items:
                        type: object
                        properties:
                          mess:
                            type: string
                            example: yuktahar
                          day:
                            $ref: "#/components/schemas/Day"
                          rate:
                            type: integer
                            example: 4210
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError" 
  /mess/capacities:
    get:
      summary: Get mess capacities for a meal
      tags:
        - Mess
      parameters:
        - name: meal
          in: query
          description: The name of the meal.
          required: true
          schema:
            $ref: "#/components/schemas/Meal"
        - name: on
          in: query
          description: Meal date for getting the capacity. Defaults to today's date.
          required: false
          schema:
            type: string
            format: date
            example: "1998-09-02"
      responses:
        '200':
          description: |
            The capacities of each mess for the given meal on the given date. Note that
            each property will be a category, and the array of items will contain each
            mess' available and maximum capacity.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    propertyNames:
                      $ref: "#/components/schemas/Category"
                    additionalProperties:
                      type: array
                      items:
                        type: object
                        properties:
                          mess:
                            type: string
                            example: yuktahar
                          available:
                            type: integer
                            example: 28
                          capacity:
                            type: integer
                            example: 420
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError" 
  /registrations:
    get:
      summary: Get registrations
      tags:
        - Registrations
      description: |
        Gets all registrations (past and upcoming) in a timeframe (max 2 months).
        Both query parameters are inclusive.
      parameters:
        - name: from
          in: query
          allowEmptyValue: false
          required: true
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: to
          in: query
          allowEmptyValue: false
          required: true
          schema:
            type: string
            format: date
            example: "2025-02-01"
      responses:
        '200':
          description: A list of registrations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MealRegistration"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError" 
    post:
      summary: Create/update a registration
      tags:
        - Registrations
      description: |
        **New in v2:** `guests` field.

        Registers the currently logged in user for a meal. If the registration
        already exists, then it updates it.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                meal_date:
                  type: string
                  format: date
                  example: "1998-09-02"
                meal_type:
                  $ref: "#/components/schemas/Meal"
                meal_mess:
                  type: string
                  example: kadamba-nonveg
                guests:
                  type: integer
                  descriptions: |
                    Number of guests this user is allowed to bring. The maximum number
                    of allowed guests is definied in the user's attributes.
                  minimum: 0
                  required: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/MealRegistration"
        '204':
          description: The registration has already been completed.
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: |
            Either the registration window has closed or the mess is
            full in capacity.
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    error:
                      allOf:
                        - $ref: "#/components/schemas/Error"
                        - type: object
                          properties:
                            code:
                              type: string
                              enum:
                                - window-closed
                                - capacity-exceeded
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError" 
  /registration:
    get:
      summary: Get one registration
      tags:
        - Registrations
      description: |
        Gets one registration for meal on date.  
        If meal is not provided, it returns registrations for all meals on that date
        in an object where key = meal and value = registration (the response body type
        you see below).  
        If date is not provided, today is assumed  
        This endpoint only returns *registered* meals.

        If a registration is missing for some meal, then 404 is returned if that meal
        is provided in the query string, otherwise, that meal's key will be missing from
        the returned object.
      parameters:
        - name: meal
          in: query
          allowEmptyValue: false
          required: false
          schema:
            $ref: "#/components/schemas/Meal"
        - name: date
          in: query
          allowEmptyValue: false
          required: false
          schema:
            type: string
            format: date
            example: "1998-09-02"
      responses:
        '200':
          description: The registration.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/MealRegistration"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError" 
...
  /registrations/cancel:
    post:
      summary: Cancel a registration
      tags:
        - Registrations
      description: |
        Cancel a registration for the meal. If the registration
        is already cancelled, then it does nothing.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                meal_date:
                  type: string
                  format: date
                  example: "1998-09-02"
                meal_type:
                  $ref: "#/components/schemas/Meal"
      responses:
        '204':
          description: Success
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: The cancellation window has closed.
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    error:
                      allOf:
                        - $ref: "#/components/schemas/Error"
                        - type: object
                          properties:
                            code:
                              type: string
                              const: window-closed
        '404':
          $ref: "#/components/responses/NotFound"
        '424':
          description: Cannot cancel since no registration exists.
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    error:
                      allOf:
                        - $ref: "#/components/schemas/Error"
                        - type: object
                          properties:
                            code:
                              type: string
                              const: not-registered
        '500':
          $ref: "#/components/responses/InternalServerError" 
  /registrations/uncancel:
    post:
      summary: Uncancel a registration
      tags:
        - Registrations
      description: |
        Uncancel a registration for the meal. If the registration
        is already uncancelled, then it does nothing.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                meal_date:
                  type: string
                  format: date
                  example: "1998-09-02"
                meal_type:
                  $ref: "#/components/schemas/Meal"
      responses:
        '204':
          description: Success
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          description: The uncancellation window has closed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    allOf:
                      - $ref: "#/components/schemas/Error"
                      - type: object
                        properties:
                          code:
                            type: string
                            const: window-closed
        '404':
          $ref: "#/components/responses/NotFound"
        '424':
          description: Cannot uncancel since registration was not cancelled.
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    error:
                      allOf:
                        - $ref: "#/components/schemas/Error"
                        - type: object
                          properties:
                            code:
                              type: string
                              const: not-cancelled
        '500':
          $ref: "#/components/responses/InternalServerError"
  /registrations/feedback:
    post:
      summary: Provide feedback for a meal
      tags:
        - Registrations
        - Feedback
      description: |
        Provide anonymous feedback for a meal. If the user is not registered
        or has not availed this meal, it returns an error.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                meal_date:
                  type: string
                  format: date
                  example: "1998-09-02"
                meal_type:
                  $ref: "#/components/schemas/Meal"
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                remarks:
                  type: string
                  required: false
      responses:
        '204':
          description: Success
        '401':
          $ref: "#/components/responses/Unauthorized"
        '409':
          description: Feedback has already been provided for this meal.
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    error:
                      allOf:
                        - $ref: "#/components/schemas/Error"
                        - type: object
                          properties:
                            code:
                              type: string
                              const: already-exists
        '424':
          description: Cannot provide feedback since meal was not availed.
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    error:
                      allOf:
                        - $ref: "#/components/schemas/Error"
                        - type: object
                          properties:
                            code:
                              type: string
                              const: not-availed
        '500':
          $ref: "#/components/responses/InternalServerError" 
  /registration/rating:
    get:
      summary: Get the average rating for a meal
      tags:
        - Registrations
        - Feedback
      description: |
        Gets the average and number of ratings for a meal in a mess
        on date. If the date is not provided, today is assumed. If
        the mess is not provided, data is returned as an object with
        the mess IDs as the keys, and the ratings objects as the
        values.

        If nobody has rated the meal, no rating(s) will be returned.
        The average rating on meals rated by users ∈ [1,5].

        If the meal's date occurs before the feedback window closes,
        the feedback will not be returned and a 403 error will be
        returned instead.
      parameters:
        - name: meal
          in: query
          allowEmptyValue: false
          required: true
          schema:
            $ref: "#/components/schemas/Meal"
        - name: mess
          in: query
          allowEmptyValue: true
          required: false
          schema:
            type: string
            example: "kadamba-nonveg"
        - name: date
          in: query
          allowEmptyValue: false
          required: false
          schema:
            type: string
            format: date
            example: "1998-09-02"
      responses:
        '200':
          description: The average rating.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schema/MealRating"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: "#/components/responses/Forbidden"
        '500':
          $ref: "#/components/responses/InternalServerError" 
...
  /registrations/cancellations:
    get:
      summary: Get number of cancelled registrations
      tags:
        - Registrations
      description: |
        Get the number of cancelled registrations in a month of a year
      parameters:
        - name: meal
          in: query
          description: The name of the meal.
          required: true
          schema:
            $ref: "#/components/schemas/Meal"
        - name: year
          in: query
          required: false
          description: The 4-digit year (2000-2098). Defaults to current year.
          schema:
            type: integer
            maximum: 2028
            minimum: 2000
        - name: month
          in: query
          required: false
          description: The month of the year (1-12). Defaults to current month.
          schema:
            type: integer
            maximum: 12
            minimum: 1
      responses:
        '200':
          description: The number.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: integer
                    maximum: 12
                    minimum: 1
                    example: 10
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError" 
  /registrations/bill:
    get:
      summary: Get the mess bill for a month
      tags:
        - Registrations
        - Billing
      description: |
        Get the mess bill for a month of a year. The bill may be projected,
        i.e. it includes bills for meals that haven't happened yet. Projected
        bills are cached, since they are expensive to calculate, but the
        cache is cleared when registrations change. A 404 is returned if
        registrations have not been opened for that month yet.
      parameters:
        - name: year
          in: query
          required: false
          description: The 4-digit year (2000-2098). Defaults to current year.
          schema:
            type: integer
            maximum: 2028
            minimum: 2000
        - name: month
          in: query
          required: false
          description: The month of the year (1-12). Defaults to current month.
          schema:
            type: integer
            maximum: 12
            minimum: 1
      responses:
        '200':
          description: The bill.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      non_projected:
                        type: integer
                        example: 1000
                        description: In paise. Non projected is the definite amount that will be charged.
                      projected:
                        type: integer
                        example: 2000
                        description: In paise. Projected is the amount that was calculated based on future registrations.
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError" 
  /extras:
    get:
      summary: List available extra items
      tags:
        - Extras
      description: |
        Lists the available extra items for a meal on a date.  
        Returns data for all messes
        If date is not provided, today is assumed.

        Some extras (with the same ID) may be available for
        multiple meals.
      parameters:
        - name: meal
          in: query
          allowEmptyValue: false
          required: true
          schema:
            $ref: "#/components/schemas/Meal"
        - name: date
          in: query
          allowEmptyValue: false
          required: false
          schema:
            type: string
            format: date
            example: "1998-09-02"
      responses:
        '200':
          description: The extra items.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExtraItem"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    Category:
      type: string
      enum:
        - registered
        - unregistered
        - guest
        - extra
    Meal:
      type: string
      enum:
        - breakfast
        - lunch
        - snacks
        - dinner
    Day:
      required: false
      type:
        - string
        - null
      enum:
        - Monday
        - Tuesday
        - Wednesday
        - Thursday
        - Friday
        - Saturday
        - Sunday
      nullable: true
    MessInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        short_name:
          type: string
        color:
          type: string
          description: |
            The color is an OKLCH value formatted as "l c h", i.e. the LCH values
            separated by a space. There is no alpha value in the color.

            To use it in CSS, you can do `color: oklch(mess.color)`,
            where `mess.color` is the mess's color.
        tags:
          $ref: "#/components/schemas/Tags"
        rating:
          type: number
          format: float
          description: The average rating of the mess, according to user feedback.
...